tgbot_server/requirements.txt
aiosqlite
alembic
fastapi[all]
SQLAlchemy
uvicorn
python-dotenv
pydantic
pyjwt
init-data-py